chore: update sampling_interval type to float for higher precision

+++ application/temperature/utils/ad5940_temperature_struct.h
@@ -44,15 +44,15 @@ AD5940_TEMPERATURE_RUN_CONFIG;
 
 typedef struct
 {
-    uint16_t sampling_interval;     /**< Sampling interval in microseconds. */
-    uint32_t TEMPSENS;              /**< Temperature sensor configuration. Refer to page 57 of the datasheet:
-                                         ```
-                                         For improved accuracy, configure the temperature sensor in chop
-                                         mode via TEMPCON0, Bits[3:1]. If chopping is selected, the user
-                                         must ensure an even number of ADC conversions take place on the
-                                         temperature sensor channel and that these results are averaged.
-                                         ```
-                                         See TEMPCON0 configuration details on page 61 of the datasheet. */
+    float sampling_interval;    /**< Sampling interval in microseconds. */
+    uint32_t TEMPSENS;          /**< Temperature sensor configuration. Refer to page 57 of the datasheet:
+                                     ```
+                                     For improved accuracy, configure the temperature sensor in chop
+                                     mode via TEMPCON0, Bits[3:1]. If chopping is selected, the user
+                                     must ensure an even number of ADC conversions take place on the
+                                     temperature sensor channel and that these results are averaged.
+                                     ```
+                                     See TEMPCON0 configuration details on page 61 of the datasheet. */
 }
 AD5940_TEMPERATURE_PARAMETERS;
